Use new ACL settings in RB

+++ migrations/versions/201505121456_45f41372799d_add_acl_settings_tables.py
@@ -6,10 +6,11 @@ Create Date: 2015-05-12 14:56:00.250271
 """
 
 import sqlalchemy as sa
-from alembic import op
+from alembic import op, context
 
 from indico.core.db.sqlalchemy import PyIntEnum
 from indico.core.db.sqlalchemy.principals import PrincipalType
+from indico.core.db.sqlalchemy.util.convert_acl_settings import json_to_acl, acl_to_json
 
 
 # revision identifiers, used by Alembic.
@@ -17,7 +18,12 @@ revision = '45f41372799d'
 down_revision = '319347ac94fe'
 
 
+acl_settings = {'roombooking.admin_principals', 'roombooking.authorized_principals'}
+
+
 def upgrade():
+    if context.is_offline_mode():
+        raise Exception('This upgrade is only possible in online mode')
     op.create_table(
         'settings_principals',
         sa.Column('type', PyIntEnum(PrincipalType), nullable=True),
@@ -95,8 +101,12 @@ def upgrade():
     op.create_index(None, 'settings_principals', ['user_id'], schema='indico')
     op.create_index(None, 'settings_principals', ['local_group_id'], schema='indico')
     op.create_index(None, 'settings_principals', ['mp_group_provider', 'mp_group_name'], schema='indico')
+    json_to_acl(acl_settings)
 
 
 def downgrade():
+    if context.is_offline_mode():
+        raise Exception('This downgrade is only possible in online mode')
+    acl_to_json(acl_settings)
     op.drop_table('settings_principals', schema='indico')
     op.drop_table('settings_principals', schema='events')
